– Remove account from account.rs – Rename account.rs to hasher.rs
