--- account.rs/src/main/index_browser_slim.ts
@@ -1,2 +1,0 @@
-export type { LightWasm, LightWasmCreator, AccountLoadOptions } from "./model.js";
-export { WasmFactory, hasSimd as hasWasmSimd } from "./wasm.js";
