--- account.rs/src/wasm/src/lib.rs
@@ -1,3 +1,0 @@
-mod account;
-mod hash;
-mod utils;
