+++ circuit-lib/circuit-lib.js/package.json
@@ -14,7 +14,7 @@
   },
   "author": "",
   "dependencies": {
-    "@lightprotocol/account.rs": "workspace:*",
+    "@lightprotocol/hasher.rs": "workspace:*",
     "@noble/curves": "^1.2.0",
     "@noble/hashes": "^1.3.2",
     "blake-hash": "^2.0.0",
