+++ cli/package.json
@@ -17,7 +17,7 @@
   ],
   "dependencies": {
     "@coral-xyz/anchor": "0.28.0",
-    "@lightprotocol/account.rs": "workspace:*",
+    "@lightprotocol/hasher.rs": "workspace:*",
     "@lightprotocol/stateless.js": "workspace:*",
     "@lightprotocol/compressed-token": "workspace:*",
     "@oclif/core": "^2.15",
