+++ hasher.rs/.gitignore
