+++ hasher.rs/benches/blake.test.ts
