+++ hasher.rs/src/main/index_browser_fat.ts
@@ -1,7 +1,8 @@
-export type { LightWasm, LightWasmCreator, AccountLoadOptions } from "./model.js";
+export type { LightWasm, LightWasmCreator, HasherLoadOptions } from "./model.js";
 export { WasmFactory, hasSimd as hasWasmSimd } from "./wasm.js";
-import wasm from "./wasm/account_wasm_bg.wasm";
-import wasmSimd from "./wasm-simd/account_wasm_simd_bg.wasm";
+
+import wasm from "./wasm/hasher_bg.wasm";
+import wasmSimd from "./wasm-simd/hasher_wasm_simd_bg.wasm";
 import { setWasmInit, setWasmSimdInit } from "./wasm.js";
 
 // @ts-ignore
