+++ hasher.rs/src/main/index_browser_slim.ts
@@ -1,0 +1,2 @@
+export type { LightWasm, LightWasmCreator, HasherLoadOptions } from "./model.js";
+export { WasmFactory, hasSimd as hasWasmSimd } from "./wasm.js";
