+++ hasher.rs/src/main/wasm.ts
@@ -1,23 +1,21 @@
 import type {
   LightWasmCreator,
-  AccountLoadOptions,
   LightWasm,
   InitInput,
   WasmInput,
+  HasherLoadOptions,
 } from "./model.js";
 
 import init, {
-  Account as AccountWasm,
   blake2 as blake2Wasm,
   blake2str as blake2strWasm,
   poseidon as poseidonWasm,
-} from "./wasm/account_wasm";
+} from "./wasm/hasher";
 import simdInit, {
-  Account as AccountSimd,
   blake2 as blake2Simd,
   blake2str as blake2strSimd,
   poseidon as poseidonSimd,
-} from "./wasm-simd/account_wasm_simd";
+} from "./wasm-simd/hasher_wasm_simd";
 import { BN } from "@coral-xyz/anchor";
 
 function stringify(input: string[] | BN[]): string[] {
@@ -39,16 +37,16 @@ export const setWasmSimdInit = (arg: () => InitInput) => {
 };
 
 const isWasmInput = (
-  x?: AccountLoadOptions["wasm"]
+  x?: HasherLoadOptions["wasm"]
 ): x is WasmInput | undefined =>
   x === undefined || (typeof x === "object" && "simd" in x);
 
 /**
- * account.rs implemented in Web assembly.
+ * hasher.rs implemented in Web assembly.
  */
 export class WasmFactory {
   static async loadModule(
-    options?: Partial<AccountLoadOptions>
+    options?: Partial<HasherLoadOptions>
   ): Promise<LightWasmCreator> {
     if (isWasmInput(options?.wasm)) {
       const useSimd = options?.simd ?? hasSimd();
@@ -62,8 +60,8 @@ export class WasmFactory {
     }
   }
 
-  static async loadAccount(
-    options?: Partial<AccountLoadOptions>
+  static async loadHasher(
+    options?: Partial<HasherLoadOptions>
   ): Promise<LightWasm> {
     const module = await WasmFactory.loadModule(options);
     return module.create();
@@ -83,27 +81,9 @@ interface HashStrategy {
   blake2str(input: string, hash_length: number): Uint8Array;
   blake2(input: Uint8Array, hash_length: number): Uint8Array;
   poseidon(inputs: Array<any>): Uint8Array;
-  seedAccount(seed: string): AccountWasm;
-  aesAccount(aes: Uint8Array): AccountWasm;
-  burnerAccount(seed: string, index: string): AccountWasm;
-  burnerSeedAccount(seed: string): AccountWasm;
-  privateKeyAccount(
-    privateKey: Uint8Array,
-    encryptionPrivateKey: Uint8Array,
-    aesSecret: Uint8Array
-  ): AccountWasm;
-  publicKeyAccount(
-    publicKey: Uint8Array,
-    encryptionPublicKey: Uint8Array | undefined
-  ): AccountWasm;
-  encryptNaclUtxo(
-    publicKey: Uint8Array,
-    message: Uint8Array,
-    commitment: Uint8Array
-  ): Uint8Array;
 }
 
-function wasmAccount(hasher: HashStrategy): LightWasmCreator {
+function wasmHasher(hasher: HashStrategy): LightWasmCreator {
   const WasmFactory = class implements LightWasm {
     blakeHash(input: string | Uint8Array, hashLength: number): Uint8Array {
       if (typeof input === "string") {
@@ -125,48 +105,6 @@ function wasmAccount(hasher: HashStrategy): LightWasmCreator {
       const bn = new BN(this.poseidonHash(input));
       return bn.toString();
     }
-
-    seedAccount(seed: string): AccountWasm {
-      return hasher.seedAccount(seed);
-    }
-
-    aesAccount(aesSecret: Uint8Array): AccountWasm {
-      return hasher.aesAccount(aesSecret);
-    }
-
-    burnerAccount(seed: string, index: string): AccountWasm {
-      return hasher.burnerAccount(seed, index);
-    }
-
-    burnerSeedAccount(seed: string): AccountWasm {
-      return hasher.burnerSeedAccount(seed);
-    }
-
-    privateKeyAccount(
-      privateKey: Uint8Array,
-      encryptionPrivateKey: Uint8Array,
-      aesSecret: Uint8Array
-    ): AccountWasm {
-      return hasher.privateKeyAccount(
-        privateKey,
-        encryptionPrivateKey,
-        aesSecret
-      );
-    }
-
-    publicKeyAccount(
-      publicKey: Uint8Array,
-      encryptionPublicKey: Uint8Array | undefined
-    ): AccountWasm {
-      return hasher.publicKeyAccount(publicKey, encryptionPublicKey);
-    }
-    encryptNaclUtxo(
-      publicKey: Uint8Array,
-      message: Uint8Array,
-      commitment: Uint8Array
-    ): Uint8Array {
-      return hasher.encryptNaclUtxo(publicKey, message, commitment);
-    }
   };
 
   return {
@@ -179,17 +117,10 @@ let simdMemory: Promise<LightWasmCreator> | undefined;
 const loadWasmSimd = async (module?: InitInput) => {
   if (simdMemory === undefined) {
     simdMemory = simdInit(module ?? wasmSimdInit?.()).then((x) => {
-      return wasmAccount({
+      return wasmHasher({
         blake2str: blake2strSimd,
         blake2: blake2Simd,
-        poseidon: poseidonSimd,
-        seedAccount: AccountSimd.new,
-        aesAccount: AccountSimd.fromAesSecret,
-        burnerAccount: AccountSimd.burner,
-        burnerSeedAccount: AccountSimd.createFromBurnerSeed,
-        privateKeyAccount: AccountSimd.fromPrivateKey,
-        publicKeyAccount: AccountSimd.fromPublicKey,
-        encryptNaclUtxo: AccountSimd.encryptNaclUtxo,
+        poseidon: poseidonSimd
       });
     });
   }
@@ -199,17 +130,10 @@ const loadWasmSimd = async (module?: InitInput) => {
 const loadWasm = async (module?: InitInput) => {
   if (sisdMemory === undefined) {
     sisdMemory = init(module ?? wasmInit?.()).then((x) => {
-      return wasmAccount({
+      return wasmHasher({
         blake2str: blake2strWasm,
         blake2: blake2Wasm,
         poseidon: poseidonWasm,
-        seedAccount: AccountSimd.new,
-        aesAccount: AccountSimd.fromAesSecret,
-        burnerAccount: AccountSimd.burner,
-        burnerSeedAccount: AccountSimd.createFromBurnerSeed,
-        privateKeyAccount: AccountSimd.fromPrivateKey,
-        publicKeyAccount: AccountSimd.fromPublicKey,
-        encryptNaclUtxo: AccountSimd.encryptNaclUtxo,
       });
     });
   }
