+++ hasher.rs/src/wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "account-wasm"
+name = "hasher"
 version = "0.1.0"
 edition = "2021"
 description = "Light Protocol Rust client"
@@ -17,18 +17,9 @@ js-sys = "0.3.64"
 web-sys = { version = "0.3.64", features = ['console'] }
 hex = { version = "0.4.3", features = [] }
 blake2b_simd = "1"
-aes-gcm-siv = "0.10.3"
-bs58 = "0.5.0"
-crypto_box = "0.7.1"
-zeroize = { version = "1.3", default-features = false }
 num-bigint = "0.4.4"
-num-traits = "0.2"
 console_error_panic_hook = "0.1.7"
-getrandom = { version = "0.2", features = ["js"] }
 thiserror = "1.0.56"
-
-# TODO: Remove once https://github.com/solana-labs/solana/issues/33504 is resolved.
-ahash = "=0.8.6"
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = false
