+++ hasher.rs/src/wasm/src/hash/blake2.rs
