+++ hasher.rs/src/wasm/src/hash/mod.rs
