+++ hasher.rs/tsconfig.json
