+++ hasher.rs/vitest.config.ts
