+++ js/compressed-token/package.json
@@ -48,7 +48,7 @@
   "dependencies": {
     "@coral-xyz/anchor": "0.28.0",
     "@lightprotocol/stateless.js": "workspace:*",
-    "@lightprotocol/account.rs": "workspace:*",
+    "@lightprotocol/hasher.rs": "workspace:*",
     "@metaplex-foundation/beet": "0.7.1",
     "@metaplex-foundation/beet-solana": "^0.4.1",
     "@solana/spl-token": "^0.3.7",
