+++ js/compressed-token/rollup.config.js
@@ -17,7 +17,7 @@ const rolls = (fmt) => ({
       '@metaplex-foundation/beet': 'beet',
       '@metaplex-foundation/beet-solana': 'beet-solana',
       '@lightprotocol/stateless.js': 'stateless.js',
-      '@lightprotocol/account.rs': 'account.rs',
+      '@lightprotocol/hasher.rs': 'hasher.rs',
       '@solana/spl-token': 'spl-token',
       buffer: 'buffer',
       crypto: 'crypto',
@@ -32,7 +32,7 @@ const rolls = (fmt) => ({
     '@coral-xyz/anchor',
     '@coral-xyz/anchor/dist/cjs/utils/bytes',
     '@lightprotocol/stateless.js',
-    '@lightprotocol/account.rs',
+    '@lightprotocol/hasher.rs',
     '@metaplex-foundation/beet',
     '@metaplex-foundation/beet-solana',
     'buffer',
