+++ js/stateless.js/package.json
@@ -63,7 +63,7 @@
   },
   "dependencies": {
     "@coral-xyz/anchor": "0.29.0",
-    "@lightprotocol/account.rs": "workspace:*",
+    "@lightprotocol/hasher.rs": "workspace:*",
     "@lightprotocol/prover.js": "workspace:*",
     "@metaplex-foundation/beet": "0.7.1",
     "@metaplex-foundation/beet-solana": "^0.4.1",
