+++ js/stateless.js/src/state/utxo-serde.ts
@@ -1,5 +1,5 @@
 import { Keypair, PublicKey } from '@solana/web3.js';
-import { LightWasm, WasmFactory } from '@lightprotocol/account.rs';
+import { LightWasm, WasmFactory } from '@lightprotocol/hasher.rs';
 import {
   Utxo,
   UtxoWithMerkleContext,
