+++ prover.js/package.json
@@ -41,7 +41,7 @@
   },
   "devDependencies": {
     "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
-    "@lightprotocol/account.rs": "workspace:*",
+    "@lightprotocol/hasher.rs": "workspace:*",
     "@lightprotocol/tsconfig": "workspace:*",
     "@rollup/plugin-typescript": "^11.1.5",
     "@solana/web3.js": "^1.87.6",
