+++ psp-template/psp-template/cargo-generate.toml
@@ -32,9 +32,9 @@ prompt = "circuit-lib.circom version"
 type = "string"
 prompt = "Prover js version"
 
-[placeholders.account-rs-version]
+[placeholders.hasher-rs-version]
 type = "string"
-prompt = "account.rs version"
+prompt = "hasher.rs version"
 
 [placeholders.zk-js-version]
 type = "string"
